+++ src/main/java/cech12/extendedmushrooms/api/recipe/ExtendedMushroomsRecipeTypes.java
@@ -1,9 +1,13 @@
 package cech12.extendedmushrooms.api.recipe;
 
+import cech12.extendedmushrooms.ExtendedMushrooms;
 import net.minecraft.item.crafting.IRecipeType;
+import net.minecraft.util.ResourceLocation;
 
 public class ExtendedMushroomsRecipeTypes {
 
-    public static IRecipeType<?> FAIRY_RING = new FairyRingRecipeType();
+    public final static ResourceLocation FAIRY_RING_ID = new ResourceLocation(ExtendedMushrooms.MOD_ID, "fairy_ring");
+
+    public static IRecipeType<?> FAIRY_RING;
 
 }